WIP
Be sure to handle deletion cascade: Product Template is created and Packaging is created. Then delete Product (this should be on Odoo to handle.)
Be sure to handle variant cascade: Product Template is created and Packaging is created. Then add Variants to the Product Template. The old Packaging is no longer associated with any Product (that Product ID has... what? been deleted?)
Be sure to handle the Packaging field on a purchase order being filled in and then cleared -- it should set the Product Unit of Measure field back to teh default but it doesn't.
Be sure that the uom dropdown under the minimal quantity field (min_qty_uom_id) only displays uoms related to the base purchase uom catetgory, or we'll get an error when we select it.
Decide if Purchase Unit of Measure (product_purchase_uom_id) should allow for Create and Edit (feels like no).
Be sure to handle the group product.group_uom (Manage Multiple Units of Measure). Maybe we should be forcing that if you are a member of the Manage Product Packaging group you are also a member of the Manage Multiple Units of Measure group and the Variant group?

- Hides the orphaned "using package" if the user is not a member of product.group_stock_packaging.